requirements.txt
arabic-reshaper==2.1.3
asgiref==3.4.1
bcrypt==3.2.0
bleach==4.1.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
dj-database-url==0.5.0
Django==3.2
django-cors-headers==3.10.0
django-heroku==0.3.1
djangorestframework==3.12.4
#drf-api-logger==1.0.8
drf-extra-fields==3.1.1
future==0.18.2
gunicorn==20.1.0
html5lib==1.1
idna==3.3
mysqlclient==2.0.3
packaging==21.3
Pillow==8.4.0
psycopg2==2.9.1
pycparser==2.21
PyJWT==2.3.0
pyparsing==3.0.6
PyPDF2==1.26.0
python-bidi==0.4.2
python-decouple==3.4
pytz==2021.1
PyYAML==5.4.1
reportlab==3.6.2
requests==2.26.0
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.7
webencodings==0.5.1
whitenoise==5.3.0
xhtml2pdf==0.2.5
